requirements.txt
fastapi
uvicorn[standard]
email-validator
sqlalchemy
asyncpg
alembic
pydantic
python-dotenv
pytest
httpx
pytest-asyncio
black
isort
